Move delete func to its own file
